功能：骰子托盘里历史记录还会记录聊天中的历史掷骰
